model/requirements.txt
keras
numpy
pillow
redis
